skin-lesion-detection-service/requirements.txt
PyJWT==1.7.1
Keras==2.2.4
marshmallow==2.16.3
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
sklearn==0.0
tensorflow==1.12.0
urllib3==1.24.1